Python script using pandas to help choose courses
Developed a freelance project back-end script in python that scrapes information from a text file about courses required to complete a CS bachelor. The program asks the user for course inputs and outputs the total amount of credits taken in a semester. Organizes the data in chart form using Panda’s python library, which first extracts the data from a classes text file. Outputs the classes taken for the semester and the number of credits intended on completing.